Refactor Python packages to match ROS standards
Created a package 'robo_magellan' for all Python packages. Moved
cone finder code to that package. Updated nodes to use new package
name for accessing the cone finder.
