+++ scripts/state_driving_to_cone.py
@@ -37,8 +37,7 @@ from state_and_transition import TRANSITION
 
 from mavros_msgs.msg import OverrideRCIn
 from robo_magellan.msg import location_msgs as Locations
-
-from cone_finder import ConeSeeker
+from robo_magellan.cone_finder import ConeSeeker
 
 # Globals
 this_node = None
