+++ setup.py
@@ -5,8 +5,7 @@ from catkin_pkg.python_setup import generate_distutils_setup
 
 # fetch values from package.xml
 setup_args = generate_distutils_setup(
-    packages=['cone_finder'],
+    packages=['robo_magellan'],
     package_dir={'': 'src'})
 
 setup(**setup_args)
-
