--- src/cone_finder/__init__.py
@@ -1,2 +1,0 @@
-""" Package initialization. Export classes to be at package level. """
-from cone_code import ConeFinder, ConeSeeker
